Fix bug in gesture recognizer
In the event of certain multi-touch events such as laying a whole hand on the iPad the below can cause a gesture to never end because the original touch appeared only a single time. This can effectively cause the entire drawing to lock up and stop updating, ignoring future touches.